feat: replaced u8 with u32 to be compatible with mysql INT UNSIGNED

+++ src/model/request_model.rs
@@ -2,18 +2,18 @@ use serde::Deserialize;
 
 #[derive(Debug, Deserialize)]
 pub struct AddItemsToTableRequest {
-    pub table_number: u8,
+    pub table_number: u32,
     pub items_names: Vec<String>,
 }
 
 #[derive(Debug, Deserialize)]
 pub struct ListTableItemsRequest {
-    pub table_number: u8,
-    pub items_ids: Option<Vec<u8>>,
+    pub table_number: u32,
+    pub items_ids: Option<Vec<u32>>,
     pub items_names: Option<Vec<String>>,
 }
 
 #[derive(Debug, Deserialize)]
 pub struct RemoveTableItemRequest {
-    pub item_id: u8,
+    pub item_id: u32,
 }
